+++ fabric-object-builder-api-v1/src/main/java/net/fabricmc/fabric/api/object/builder/v1/block/FabricBlockMaterial.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package net.fabricmc.fabric.api.object.builder.v1;
+package net.fabricmc.fabric.api.object.builder.v1.block;
 
 import com.google.common.annotations.Beta;
